+++ src/test/java/com/cloudant/tests/ChangeNotificationsTest.java
@@ -54,7 +54,7 @@ public class ChangeNotificationsTest {
 
 	@After
 	public void tearDown(){
-		account.deleteDB("lightcouch-db-test", "delete database");
+		account.deleteDB("lightcouch-db-test");
 		account.shutdown();
 	}