+++ src/test/java/com/cloudant/tests/DatabaseTest.java
@@ -43,7 +43,7 @@ public class DatabaseTest {
 
 	@After
 	public  void tearDown() {
-		account.deleteDB("animaldb", "delete database");
+		account.deleteDB("animaldb");
 		account.shutdown();
 	}