+++ src/test/java/com/cloudant/tests/SearchTests.java
@@ -52,7 +52,7 @@ public class SearchTests {
 
 	@After
 	public  void tearDown() {
-		account.deleteDB("animaldb", "delete database");
+		account.deleteDB("animaldb");
 		account.shutdown();
 	}